6) Математические операции - Сложение Float

+++ src/com/company/Main.java
@@ -6,11 +6,11 @@ public class Main {
 
     public static void main(String[] args) {
         Scanner num = new Scanner(System.in);
-        int first, second, result;
+        float first, second, result;
         System.out.println("Enter first num: ");
-        first = num.nextInt();
+        first = num.nextFloat();
         System.out.println("Enter second num");
-        second = num.nextInt();
+        second = num.nextFloat();
         result = first + second;
         System.out.println("Result is: " + result);
     }
